+++ playtimestats-plugin-core/src/main/kotlin/io/github/arcaneplugins/playtimestats/plugin/core/data/PlaytimeData.kt
@@ -24,6 +24,6 @@ import java.util.*
 data class PlaytimeData(
     val uuid: UUID,
     val lastUsername: String,
-    val minutesPlayed: Int,
+    val minutesPlayed: Float,
     val sessionsPlayed: Int,
 )
